projekt/requirements.txt
argparse
json
xmltodict
PyYAML